pyproject.toml
[tool.poetry]
name = "htmx-demo"
version = "0.1.0"
description = ""
authors = ["boggs <hello@boggs.xyz>"]
readme = "README.md"
packages = [{include = "htmx_demo"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.5"
django-extensions = "^3.2.3"
gunicorn = "^21.2.0"
dj-database-url = "^2.1.0"
whitenoise = "^6.5.0"
psycopg = "^3.1.10"
sentry-sdk = "^1.30.0"
factory-boy = "^3.3.0"
django-htmx = "^1.16.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
ruff = "^0.0.286"
pre-commit = "^3.4.0"
bandit = "^1.7.5"
detect-secrets = "^1.4.0"
djhtml = "^3.0.6"
dslr = {version="^0.4.0", extras=["psycopg2-binary"]}
django-types = "^0.17.0"
pyright = "^1.1.326"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "F", "W", "C90", "I"]

[tool.ruff.per-file-ignores]
"htmx_demo/*/migrations/*.py" = ["E501"]

[tool.bandit]
exclude_dirs = ["tests"]
